+++ src/Renderer/renderer.hpp
@@ -18,8 +18,8 @@ public:
 	
 	static void beginScene(Camera&);
 
-	static void submit(VertexArray& vao, Shader& shader, const sf::Transform&, sf::Vector2f size);
-	static void submit(VertexArray& vao, const sf::Transform&, sf::Vector2f size);
+	static void submit(VertexArray& vao, Shader& shader, const sf::Transform&, sf::Vector2i size);
+	static void submit(VertexArray& vao, const sf::Transform&, sf::Vector2i size);
 	static void submit(Sprite&, Shader&, const sf::Transform&);
 	static void submit(Sprite&, const sf::Transform&);
